Splits client/server max_window_bits
