+++ examples/autobahn-server.rs
@@ -2,7 +2,8 @@ use std::net::{TcpListener, TcpStream};
 use std::thread::spawn;
 
 use log::*;
-use tungstenite::extensions::deflate::{DeflateExt, DeflateConfigBuilder};
+use tungstenite::extensions::compression::deflate::DeflateConfigBuilder;
+use tungstenite::extensions::compression::WsCompression;
 use tungstenite::handshake::HandshakeRole;
 use tungstenite::protocol::WebSocketConfig;
 use tungstenite::server::accept_with_config;
@@ -25,7 +26,7 @@ fn handle_client(stream: TcpStream) -> Result<()> {
         Some(WebSocketConfig {
             max_send_queue: None,
             max_frame_size: Some(16 << 20),
-            encoder: DeflateExt::new(deflate_config),
+            compression: WsCompression::Deflate(deflate_config),
         }),
     )
     .map_err(must_not_block)?;
