+++ src/handshake/server.rs
@@ -12,7 +12,7 @@ use super::headers::{FromHttparse, MAX_HEADERS};
 use super::machine::{HandshakeMachine, StageResult, TryParse};
 use super::{convert_key, HandshakeRole, MidHandshake, ProcessingResult};
 use crate::error::{Error, Result};
-use crate::extensions::verify_compression_req_headers;
+use crate::extensions::compression::verify_compression_req_headers;
 use crate::protocol::{Role, WebSocket, WebSocketConfig};
 
 /// Server request type.
